+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Open UI preview of transport graph when `lattice.draw()` is invoked and print URL of the same
 - Defer creation of server specific config entries until covalent is started
 - Functional tests on CI
+- Move dask worker space to covalent cache instead of `site-packages`
 
 ### Docs